libnixf/Basic: split basic nodes headers (#348)
