+++ libnixf/src/Parse/ParseAttrs.cpp
@@ -1,4 +1,6 @@
-#include "Parser.h"
+#include "ParserImpl.h"
+
+#include "nixf/Basic/Nodes/Attrs.h"
 
 namespace nixf {
