+++ libnixf/src/Parse/ParseLambda.cpp
@@ -1,4 +1,6 @@
-#include "Parser.h"
+#include "ParserImpl.h"
+
+#include "nixf/Basic/Nodes/Lambda.h"
 
 namespace nixf {
